requirements/base.txt
Django==1.7.7
psycopg2>=2.5.4
django-compressor>=1.4
django-modelcluster>=0.4
wagtail==1.0b1
django-taggit>=0.12.2
django-libsass>=0.2
beautifulsoup4==4.3.2
wagtailsweetcaptcha==0.0.1